backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pymongo==4.6.1
load_dotenv==0.1.0
